Some initial modifications
Made it look more like a standard Rack/Sinatra project, i.e. Sinatra is not included in the top-level, the app is run with config.ru, there is a Gemfile etc. Note that I called the Sinatra::Application object BunkerNews, rather than LamerNews, since the idea is to try to deploy something we can use in and around the Bunker.